Set breakpoint in a file not loaded yet.
Modified PageAgent.getResourceTree() to guess what is the application's root
directory based on path in process's argv[1] and to list all .js files under
this root directory.

Note that global modules and node.js internals are not listed.

Also modified behaviour of the application, we use argv1 file as url for
the top-level frame. This way we can always load resource content from
disk (we don't have to try to load it from V8 first), since getResourceTree
lists only existing files now.

This commit allows users to set breakpoints in files that are not loaded yet,
which is helpful for debugging module-loading problems.