Handle shebang in ScriptFileStorage load/save
Before this change, ScriptFileStorage.load was returning source code
including shebang (#!/usr/bin/node). This prevented the front-end
from matching the file on disk with the script loaded in Node/V8.
As a result, `node-debug _mocha` did not show the breakpoint in
the _mocha file.

ScriptFileStorage.save is preserving shebang when saving the updated
file now, so that live edit works as expected.