BreakEventHandler: fix break on uncaught exception
Fix a bug where `TypeError: undefined is not a function` was ignored
in "break on uncaught exceptions" mode.

This exception is thrown by a V8 builtin function CALL_NON_FUNCTION
defined in an internal script `v8/src/runtime.js`, which is not
tracked by Node Inspector. Before this change, Node Inspector
incorrectly assumed that the script is a code loaded via `eval`.