Fix issue #832, "node inspector and babel-register".
This fixes the issue on Mac OSX where it appears to be impossible to
enable any breakpoints while debugging ES6 code dynamically transpiled
by either babel-node or babel-register.
The initial cause of the problem is that the source and the target file
names in source maps produced by babel-register are exactly equal and
have no paths.
node-inspector already contains code that addresses this problem:
_checkSourceMapIssues()/fixWrongFileName() in lib/ScriptManager.js,
which appends ".source" extension to the source file name in the source
map, when name duplication is detected. Unfortunately the code converting
URLs to absolute paths in that method does not work correctly with
Linux/OSX URLs: it strips the root slash from the path thus rendering
path invalid. This commit fixes this bug and also, exactly same bug
in lib/ScriptManager.normalizeName().